Separate GUI into separate class Gui
