+++ src/main/java/duke/DukeLauncher.java
@@ -7,6 +7,6 @@ import javafx.application.Application;
  */
 public class DukeLauncher {
     public static void main(String[] args) {
-        Application.launch(Duke.class, args);
+        Application.launch(Gui.class, args);
     }
 }
